Create LSB BASED IMAGE AND TEXT STEGANOGRAPHY WITH ENHANCED SECURITY USING TRIPLE DES
LSB-based image steganography with 3DES encryption is to provide a robust 
and secure platform for embedding and extracting data within digital images. This 
system encompasses multiple key components, including a user-friendly interface for 
image and data selection, a 3DES encryption module to secure the data, and an LSB 
embedding module for discreetly incorporating the encrypted information into a 
chosen cover image. Key management plays a critical role, ensuring the security of 
both encryption and steganography keys. The system generates stego images, which 
can be viewed and exported. To extract the concealed data, users with the appropriate 
keys can reverse the process, revealing the hidden content. Security measures, such 
as user authentication and access control, are implemented to safeguard the system. 
The following steps are for processing steaganography with 3DES:
First the secret message is encrypted using 3DES symmetric algorithm 
with the key given by user. 
The given key is converted into bytes using MD5 algorithm.  
Then the cipher text is converted into binary using ASCII values.  
Now changing the LSB values of image pixels based on ASCII value of 
cipher.  
Then the cipher message is hidden in image using LSB algorithm.   
Encoded image is formed.  
For decoding, same LSB method is used in reverse process to extract the 
hidden data.  
Decoded text is extracted from the image.  
The extracted data is cipher text, so the same 3DES symmetric key used 
to decrypt the message.  
Finally, the secret message is obtained.