Create SMART BRIDGE AUTOMATION FOR FLOOD  USING IOT
An automatic height adjusting bridge is designed to maintain safe height during 
heavy rains or floods. A smart bridge is one that senses some significant condition of 
water level and then automatically increases its height to ensure safety to the vehicles 
and pedestrians .Bridges are essential infrastructure that connects different areas and 
makes transportation more accessible. 
However, they can be challenging to maintain, especially when water levels increase 
due to heavy rainfall or floods. In such cases, bridges can become dangerous, causing 
traffic to come to a halt or even collapse. To prevent this, engineers have developed 
an automatic height-adjusting bridge that can helps to maintain the safety of the bridge 
even during heavy rain or floods. This bridge is equipped with an Arduino, servo 
motor, moisture sensor, and other components that help adjust its height based on the 
water level.